requirements.txt
flask
transformers
torch
sentence-transformers
python-docx
pandas
openpyxl
python-pptx
PyPDF2
